recebendo arquivos de 3G

+++ src/include/client_utils.h
@@ -13,8 +13,8 @@
 
 char *build_file_path(char* file_name, char *dir_path);
 int create_connection(char *host, char *port);
-void parse_header(int con_sock, int *num_threads, int *file_size);
-void server_thread_params(int con_sock, int *offset, int *segment_size);
+void parse_header(int con_sock, int *num_threads, long *file_size);
+void server_thread_params(int con_sock, long *offset, long *segment_size);
 void read_line(int fd, char conent[]);
 
 #endif /* CLIENT_UTILS_H_ */
